requirements.txt
BeautifulSoup
GitPython==2.1.3
threadpool